sdk/testing: move reloadutil into internal package
