+++ sdk/testing/stepwise/environments/docker/environment.go
@@ -35,7 +35,7 @@ import (
 	uuid "github.com/hashicorp/go-uuid"
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/sdk/helper/consts"
-	"github.com/hashicorp/vault/sdk/helper/reloadutil"
+	"github.com/hashicorp/vault/sdk/testing/internal/reloadutil"
 	"github.com/hashicorp/vault/sdk/testing/stepwise"
 	"golang.org/x/net/http2"
 )
